// ChatSX/notifications/models/chatroomData.ts
export interface ChatroomData {
  chatroomId: number;
  chatroomLastConversation: string;
  chatroomLastConversationTimestamp: number;
  chatroomLastConversationUserImage: string;
  chatroomLastConversationUserName: string;
  chatroomName: string;
  chatroomTitle: string;
  chatroomUnreadConversationCount: number;
  chatroomUserImage: string;
  chatroomUserName: string;
  communityId: string;
  communityImage: string;
  communityName: string;
  notificationId: string;
  route: string;
  routeChild: string;
}